lib: Print OSC 7 escape sequence (current host and working directory) for many Unix terminal emulators
Previously this was only emitted on on macOS with Apple's iTerm.app (and
compatible clones like iTerm2), but it is used by other terminal emulators
as well to obtain the actual current working directory wiht symlinks intact.
All non-supporting terminal emulators tested gracefully ignored this value,
so emit this on these as well in case they (or some other app masquarading
as them) add future support for this value.